% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_stage_cov.R
\name{final_stage_cov}
\alias{final_stage_cov}
\title{Computes the sample size covariance matrix given stage sample size and observed frequency. This only works for one strata. For multiple stratum, this function is applied recursively, assuming independence between stratum.}
\usage{
final_stage_cov(n_list, p)
}
\arguments{
\item{n_list}{Integer, sample size at each stage. Each Each element denotes the sample size of a stage. The number of elements denote the number of stages.}

\item{p}{Numeric, vector of observed frequencies, should add up to 1.}
}
\value{
A list. The first element is a numeric matrix, which is the sample size variance covariance matrix. The second element is numeric, its the equivalent sample size across stages. The third element is a numeric matrix, which is the the unmodified frequency variance covariance.
}
\description{
Computes the sample size covariance matrix given stage sample size and observed frequency. This only works for one strata. For multiple stratum, this function is applied recursively, assuming independence between stratum.
}
