% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstagetest.R
\name{mstagetest}
\alias{mstagetest}
\title{Homogeneity test for multi-stage sampling.}
\usage{
mstagetest(n0, n1, f0, f1)
}
\arguments{
\item{n0}{Numeric vector/matrix. The sample size of population 0. For one stratum, a numeric vector, with each element being the sample size of the stage. For multiple strata, a matrix, with the rows being the strata and columns the stages.}

\item{n1}{Numeric vector/matrix. The sample size of population 1. Same form as n0.}

\item{f0}{Numeric vector/matrix. Observed frequencies of population 0. For one stratum, a numeric vector. For multiple strata, a matrix with the rows being the strata and columns the frequencies.}

\item{f1}{Numeric vector/matrix. Observed frequencies of population 1. Same form as f0.}
}
\value{
A list with three elements. They are p-values from three different approaches. The first is the approach where there is no frequency pooling between strata/populations. The second with pooling between populations. The third with pooling between strata and populations.
}
\description{
The test must be between two populations, but the number of stages does not need to be the same for the two populations.
}
