% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_stages.R
\name{simulate_stages}
\alias{simulate_stages}
\title{Given starting frequencies, simulate frequencies after some stages using multinomial sampling.}
\usage{
simulate_stages(seq_counts, p1, seed = NULL)
}
\arguments{
\item{seq_counts}{Integer, sample size at each stage. Each Each element denotes the sample size of a stage. The number of elements denote the number of stages.}

\item{p1}{Numeric, vector of starting frequencies. Should add up to 1.}

\item{seed}{Numeric, simulate using the given seed. Will not have a fix seed with no input.}
}
\value{
A list. First element is the simulated frequency. Second element the sample size of the last stage.
}
\description{
Given starting frequencies, simulate frequencies after some stages using multinomial sampling.
}
